refactor(ui): extract transcript scroll margin
