+++ ui/src/pages/chat/components/chat-transcript-scroll-margin.ts
@@ -1,0 +1,30 @@
+import type { Virtualizer } from "@tanstack/virtual-core";
+import type { ReactiveControllerHost } from "lit";
+
+function transcriptScrollMargin(element: Element | null): number {
+  if (!(element instanceof HTMLElement) || typeof getComputedStyle !== "function") {
+    return 0;
+  }
+  const margin = Number.parseFloat(getComputedStyle(element).paddingTop);
+  return Number.isFinite(margin) ? margin : 0;
+}
+
+export function initialChatTranscriptScrollMargin(host: ReactiveControllerHost): number {
+  return host instanceof HTMLElement
+    ? transcriptScrollMargin(host.querySelector(".chat-thread"))
+    : 0;
+}
+
+export function syncChatTranscriptScrollMargin(
+  scrollElement: HTMLDivElement | null,
+  virtualizer: Virtualizer<HTMLDivElement, HTMLElement>,
+): void {
+  const scrollMargin = transcriptScrollMargin(scrollElement);
+  if (scrollMargin === virtualizer.options.scrollMargin) {
+    return;
+  }
+  virtualizer.setOptions({
+    ...virtualizer.options,
+    scrollMargin,
+  });
+}
